Added negligible delay before adding/deleting tags
When testing TagManager, I ran into an error where the name history
HashMap entries were getting overwritten because the timestamps of
actions were too close together. So, I added a 1ms delay before each
action to prevent this from happening. Saad helped me debug this error.

A better implementation of storing the name history would be to simply
use a List, but due to the approaching due date, we decided to simply
add in this hotfix for now instead of refactoring a lot of our code.