Fixed infinite loop bug
When adding to the available tag set, you could get stuck in an
infinite loop if you didn't enter a valid tag name. So, you couldn't
simply exit from the first popup window without adding a tag. I fixed
it so that the user can simply exit from the text input dialog without
being forced to enter a new tag.

Another thing I changed was that I replaced many instances of the
hard-coded name of the placeholder image with a static final variable
in the ImageTagManager class.

Finally, I updated the class for resetting test files to delete the log
files in addition to the ser file.